Separate corefx-specific tests
