+++ src/System.IO.FileSystem/tests/File/Move.cs
@@ -7,32 +7,13 @@ using System.Linq;
 
 namespace System.IO.Tests
 {
-    public class File_Move : FileSystemTest
+    public partial class File_Move : FileSystemTest
     {
         #region Utilities
 
         public virtual void Move(string sourceFile, string destFile)
         {
             File.Move(sourceFile, destFile);
-        }
-
-        public virtual void Move(string sourceFile, string destFile, bool overwrite)
-        {
-            File.Move(sourceFile, destFile, overwrite);
-        }
-
-        private void MoveDestinationFileDoesNotExist(bool overwrite)
-        {
-            string srcPath = GetTestFilePath();
-            string destPath = GetTestFilePath();
-
-            byte[] srcContents = new byte[] { 1, 2, 3, 4, 5 };
-            File.WriteAllBytes(srcPath, srcContents);
-
-            Move(srcPath, destPath, overwrite);
-
-            Assert.False(File.Exists(srcPath));
-            Assert.Equal(srcContents, File.ReadAllBytes(destPath));
         }
 
         #endregion
@@ -110,35 +91,6 @@ namespace System.IO.Tests
         }
 
         [Fact]
-        public void BasicMoveWithOverwriteFileExists()
-        {
-            string srcPath = GetTestFilePath();
-            string destPath = GetTestFilePath();
-
-            byte[] srcContents = new byte[] { 1, 2, 3, 4, 5 };
-            byte[] destContents = new byte[] { 6, 7, 8, 9, 10 };
-            File.WriteAllBytes(srcPath, srcContents);
-            File.WriteAllBytes(destPath, destContents);
-
-            Move(srcPath, destPath, true);
-
-            Assert.False(File.Exists(srcPath));
-            Assert.Equal(srcContents, File.ReadAllBytes(destPath));
-        }
-
-        [Fact]
-        public void BasicMoveWithOverwriteFileDoesNotExist()
-        {
-            MoveDestinationFileDoesNotExist(true);
-        }
-
-        [Fact]
-        public void BasicMoveWithoutOverwriteFileDoesNotExist()
-        {
-            MoveDestinationFileDoesNotExist(false);
-        }
-
-        [Fact]
         public void MoveNonEmptyFile()
         {
             FileInfo testFileSource = new FileInfo(GetTestFilePath());
@@ -174,23 +126,6 @@ namespace System.IO.Tests
             Assert.Throws<IOException>(() => Move(testFileSource.FullName, testFileDest.FullName));
             Assert.True(File.Exists(testFileSource.FullName));
             Assert.True(File.Exists(testFileDest.FullName));
-        }
-
-        [Fact]
-        public void MoveOntoExistingFileNoOverwrite()
-        {
-            string srcPath = GetTestFilePath();
-            string destPath = GetTestFilePath();
-
-            byte[] srcContents = new byte[] { 1, 2, 3, 4, 5 };
-            byte[] destContents = new byte[] { 6, 7, 8, 9, 10 };
-            File.WriteAllBytes(srcPath, srcContents);
-            File.WriteAllBytes(destPath, destContents);
-
-            Assert.Throws<IOException>(() => Move(srcPath, destPath, false));
-            Assert.True(File.Exists(srcPath));
-            Assert.True(File.Exists(destPath));
-            Assert.Equal(destContents, File.ReadAllBytes(destPath));
         }
 
         [Fact]
