+++ src/System.IO.FileSystem/tests/System.IO.FileSystem.Tests.csproj
@@ -16,6 +16,7 @@
     <Compile Include="Directory\EnumerableTests.cs" />
     <Compile Include="FileInfo\GetSetAttributesCommon.cs" />
     <Compile Include="FileInfo\IsReadOnly.cs" />
+    <Compile Include="FileInfo\MoveTo.netcoreapp.cs" />
     <Compile Include="FileInfo\Replace.cs" />
     <Compile Include="FileStream\Handle.cs" />
     <Compile Include="Directory\GetLogicalDrives.cs" />
@@ -25,6 +26,7 @@
     <Compile Include="FileSystemTest.cs" />
     <Compile Include="File\EncryptDecrypt.cs" />
     <Compile Include="File\GetSetAttributesCommon.cs" />
+    <Compile Include="File\Move.netcoreapp.cs" />
     <Compile Include="File\Replace.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetsUnix)' == 'true'">
